# Conflicts:
#	GlyfiBot/Services/DuplicateMessageCleanerService.cs

+++ GlyfiBot/Services/DuplicateMessageCleanerService.cs
@@ -1,3 +1,4 @@
+using GlyfiBot.Commands;
 using NetCord;
 using NetCord.Gateway;
 using NetCord.Rest;
@@ -74,6 +75,19 @@ public static class DuplicateMessageCleanerService
 				}
 				return result.ToString();
 			}
+		}
+
+		public async Task<bool> Exists()
+		{
+			try
+			{
+				await _client.Rest.GetMessageAsync(ChannelId, Id);
+			}
+			catch(RestException restException) when(restException.StatusCode == HttpStatusCode.NotFound)
+			{
+				return false;
+			}
+			return true;
 		}
 
 		public bool IsEqualTo(UserMessage other)
@@ -171,20 +185,24 @@ public static class DuplicateMessageCleanerService
 					// ...then we compare contents (including attachments!)
 					if (thisMessage.IsEqualTo(prevMessage))
 					{
-						// If they are the same, then we stop the spam!
+						// If the previous message has been deleted, then we don't do anything.
+						if (await prevMessage.Exists())
+						{
+							// The messages the same! Stop the spam!
 
-						// First we timeout (to prevent further infractions) and we let the mods know
-						await Task.WhenAll([
-							TimeoutUser(guildUser),
-							NotifyMods(prevMessage, thisMessage),
-						]);
+							// First we timeout (to prevent further infractions) and we let the mods know
+							await Task.WhenAll([
+								TimeoutUser(guildUser),
+								NotifyMods(prevMessage, thisMessage),
+							]);
 
-						// Lastly, we clean up the mess
-						await Task.WhenAll([
-							NotifyUserOfTimeout(guildUser),
-							DeleteMessageIfExists(prevMessage),
-							DeleteMessageIfExists(thisMessage),
-						]);
+							// Lastly, we clean up the mess
+							await Task.WhenAll([
+								NotifyUserOfTimeout(guildUser),
+								DeleteMessageIfExists(prevMessage),
+								DeleteMessageIfExists(thisMessage),
+							]);
+						}
 					}
 				}
 			}
@@ -325,11 +343,14 @@ public static class DuplicateMessageCleanerService
 				MessageReference = MessageReferenceProperties.Forward(thisMessage.ChannelId, thisMessage.Id),
 			});
 
+			string? modsPing = SetModPingCommand.GetModsPing(guildId.Value);
+			modsPing = modsPing is null ? "" : $", {modsPing}";
+
 			await _client.Rest.SendMessageAsync(channelId, new MessageProperties
 			{
 				Components =
 				[
-					new TextDisplayProperties($"{prevMessage.Author} sent this↑ message in {prevMessage.Channel} and {thisMessage.Channel}!"),
+					new TextDisplayProperties($"{prevMessage.Author} sent this↑ message in {prevMessage.Channel} and {thisMessage.Channel}{modsPing}!"),
 					new TextDisplayProperties($"The messages have been cleaned up, and the account has been given a timeout of {TIMEOUT_TIME_MINUTES} minutes."),
 					new ActionRowProperties([
 						new ButtonProperties(new InteractionDataContainer<ulong>(
